mnist/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anyio=3.5.0=py38h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=py38h06a4308_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py38h06a4308_0
black=22.6.0=py38h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py38h7deecbd_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.9.24=py38h06a4308_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
contourpy=1.0.6=pypi_0
cryptography=38.0.1=py38h9ce1e76_0
cuda=11.6.2=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=11.8.86=0
cuda-libraries=11.6.2=0
cuda-libraries-dev=11.6.2=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=11.8.87=0
cuda-runtime=11.6.2=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.6.2=0
cuda-tools=11.6.2=0
cuda-visual-tools=11.6.2=0
cycler=0.11.0=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
gds-tools=1.4.0.31=0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.4=py38h06a4308_0
importlib-metadata=4.11.3=py38h06a4308_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.2=py38h06a4308_0
ipython=8.7.0=py38h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py38h06a4308_1
jinja2=3.1.2=py38h06a4308_0
joblib=1.1.1=py38h06a4308_0
jpeg=9e=h7f8727e_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.16.0=py38h06a4308_0
jupyter=1.0.0=py38h06a4308_8
jupyter_client=7.4.7=py38h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.11.2=py38h06a4308_0
jupyter_server=1.18.1=py38h06a4308_0
jupyterlab=3.5.0=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.16.3=py38h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.4=pypi_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=10.0.1=default_hb85057a_2
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=hbcb73fb_5
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lxml=4.9.1=py38h1edc446_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.6=py38h06a4308_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mypy_extensions=0.4.3=py38h06a4308_1
nbclassic=0.4.8=py38h06a4308_0
nbclient=0.5.13=py38h06a4308_0
nbconvert=6.5.4=py38h06a4308_0
nbformat=5.7.0=py38h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py38h06a4308_0
nettle=3.7.3=hbbd107a_1
notebook=6.5.2=py38h06a4308_0
notebook-shim=0.2.2=py38h06a4308_0
nsight-compute=2022.3.0.22=0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py38he184ba9_0
numpy=1.23.4=py38h14f4228_0
numpy-base=1.23.4=py38h31eccc5_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.2=py38h417a72b_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathspec=0.9.0=py38h06a4308_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
pkgutil-resolve-name=1.3.10=py38h06a4308_0
platformdirs=2.5.2=py38h06a4308_0
ply=3.11=py38_0
prometheus_client=0.14.1=py38h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-sip=12.11.0=py38h6a678d5_1
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py38h06a4308_0
pytorch=1.13.0=py3.8_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_0
pytorch-mutex=1.0=cuda
pytz=2022.6=pypi_0
pyzmq=23.2.0=py38h6a678d5_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtconsole=5.3.2=py38h06a4308_0
qtpy=2.2.0=py38h06a4308_0
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.1=py38h06a4308_0
scikit-learn=1.1.3=py38h6a678d5_0
scipy=1.9.3=py38h14f4228_0
seaborn=0.12.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.5.0=py38h06a4308_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py38h06a4308_1
soupsieve=2.3.2.post1=py38h06a4308_0
sqlite=3.40.0=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.13.1=py38h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py38h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py38h06a4308_0
torchaudio=0.13.0=py38_cu116
torchvision=0.14.0=py38_cu116
tornado=6.2=py38h5eee18b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.4.0=py38h06a4308_0
typing_extensions=4.4.0=py38h06a4308_0
urllib3=1.26.13=py38h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=py38h06a4308_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py38h06a4308_0
xz=5.2.8=h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.8.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
